% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjmat.r
\name{adjmat}
\alias{adjmat}
\title{Generate a sparse adjacency matrix}
\usage{
adjmat(x)
}
\arguments{
\item{x}{The n x p design matrix.}
}
\value{
A list
\itemize{
\item edge - The nonzero elements of the adjacency matrix.
\item edgerow - The row index of each nonzero elements, which has the same length as \code{edge}.
\item edgecol - The column index of each nonzero elements, which has the same length as \code{edge}.
\item gamma - A \code{p} vector. Each elements is the sum of the column of the adjacency matrix.
\item ledge - The number of the nonzero elements.
}
}
\description{
This function generates a sparse adjacency matrix for predictors.
The elements of the matrix indicate whether pairs of vertices are adjacent or not in the graph.
And each elements is based on the Pearson's correlation coefficients and proper cut-off value.
To save the storage space, this function only retains the information of nonzero ones
}
\examples{
x = matrix(rnorm(200), 10, 20)
A <- adjmat(x)
}
\references{
Jian Huang. Shuangge Ma. Hongzhe Li and Cun-Hui Zhang. (2011)
\emph{The sparse Laplacian shrinkage estimator for high-dimensional regression}
}
