% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.r
\name{generator}
\alias{generator}
\title{Generate survival samples}
\usage{
generator(
  n,
  p,
  d,
  g,
  sig,
  rho,
  error = c("norm", "contaminate", "t2", "ev"),
  tran = c("lm", "log"),
  censor.rate = 0,
  block = c("Block", "Auto", "BAND")
)
}
\arguments{
\item{n}{The number of samples.}

\item{p}{The number of covariates.}

\item{d}{The number of nonzero values of true coefficients.}

\item{g}{The group size.}

\item{sig}{The two parameters of the uniform distribution. For example, \code{sig=c(a,b)}, which means the elements of true coefficients are from \code{U(a,b)}.}

\item{rho}{The strength of correlation among covariates.}

\item{error}{The error distribution. Quantitative for error="norm", error="contaminate", error="t2" or error="ev". For error="contaminate", y is from \code{0.7N(0,1)+0.3Cauchy}.}

\item{tran}{The transformation function. Quantitative for tran="lm" or tran="log"}

\item{censor.rate}{The censor rate of survival sample. Default is 0.}

\item{block}{The network structure of predictors. Quantitative for block="Block", block="Auto", block="BAND". For block="BAND", the number of groups \code{g} is equal to \code{p}.}
}
\value{
A list
\itemize{
\item x - The \code{n x p} design matrix.
\item y - The \code{n} repsonse.
\item b - The true coefficients.
\item status - The survival status. \code{1} is censored and \code{0} is uncensored.
}
}
\description{
This function generates survival data including response, covariates and status.
Data can be right-censored and heavy-tailed.
Morever, there may be a strong network structure and high-correlated pattern among covariates.
}
\examples{
library(mvtnorm)
library(Matrix)

n = 400
p = 500
d = 15
g = 5
sig = c(0.5, 1.5)
rho = 0.9
error = "contaminate"
tran = "log"
censor.rate = 0.1
block = "Auto"

dat = generator(n, p, d, g, sig, rho, error, tran, censor.rate, block)
x = dat$x
y = dat$y
b = dat$b
status = dat$status
}
